migrated all redux functionality to redux context
