+++ src/containers/LeftPanel/ActionCreator/ActionCreator.jsx
@@ -6,7 +6,7 @@ import {
   updateActionCreator,
   updateActionsFilePath,
   updateTypesFilePath,
-} from '../../../context/testCaseActions';
+} from '../../../context/reduxTestCaseActions';
 import { Draggable } from 'react-beautiful-dnd';
 const closeIcon = require('../../../assets/images/close.png');
 const dragIcon = require('../../../assets/images/drag-vertical.png');
