+++ src/containers/LeftPanel/Reducer/Reducer.jsx
@@ -6,7 +6,7 @@ import {
   updateReducer,
   updateTypesFilePath,
   updateReducersFilePath,
-} from '../../../context/testCaseActions';
+} from '../../../context/reduxTestCaseActions';
 import { Draggable } from 'react-beautiful-dnd';
 const closeIcon = require('../../../assets/images/close.png');
 const dragIcon = require('../../../assets/images/drag-vertical.png');
@@ -112,10 +112,9 @@ const Reducer = ({ reducer, index, dispatchToTestCase }) => {
             </div>
           </div>
         </div>
-      )
-      }
+      )}
     </Draggable>
-  )
+  );
 };
 
 export default Reducer;
