+++ src/containers/LeftPanel/TestCase/TestStatements.jsx
@@ -2,35 +2,13 @@ import React from 'react';
 import Action from '../Action/Action';
 import Assertion from '../Assertion/Assertion';
 import Rerender from '../Render/Rerender';
-import Middleware from '../Middleware/Middleware';
 import Context from '../Context/Context';
-import ActionCreator from '../ActionCreator/ActionCreator';
-import Async from '../Thunk/Thunk';
-import Reducer from '../Reducer/Reducer';
-import HookRender from '../HookRender/HookRender'
+import HookRender from '../HookRender/HookRender';
 import HookUpdates from '../HookUpdates/HookUpdates';
 
 const TestStatements = function TestStatements({ statements, dispatchToTestCase }) {
   return statements.map((statement, i) => {
     switch (statement.type) {
-      case 'middleware':
-        return (
-          <Middleware
-            key={statement.id}
-            middleware={statement}
-            index={i}
-            dispatchToTestCase={dispatchToTestCase}
-          />
-        );
-        case 'context':
-          return ( 
-            <Context
-              key={statement.id}
-              context={statement}
-              index={i}
-              dispatchToTestCase={dispatchToTestCase}
-            />
-          );
       case 'action':
         return (
           <Action
@@ -49,32 +27,15 @@ const TestStatements = function TestStatements({ statements, dispatchToTestCase 
             dispatchToTestCase={dispatchToTestCase}
           />
         );
-      // reducer
-      case 'reducer':
-        return (
-          <Reducer
-            key={statement.id}
-            reducer={statement}
-            index={i}
-            dispatchToTestCase={dispatchToTestCase}
-          />
-        );
-      case 'async':
-        return (
-          <Async
-            key={statement.id}
-            async={statement}
-            index={i}
-            dispatchToTestCase={dispatchToTestCase}
-          />
-        );
+
       case 'render':
         return <Rerender key={statement.id} render={statement} index={i} />;
-      case 'action-creator':
+
+      case 'context':
         return (
-          <ActionCreator
+          <Context
             key={statement.id}
-            actionCreator={statement}
+            context={statement}
             index={i}
             dispatchToTestCase={dispatchToTestCase}
           />
