+++ src/containers/LeftPanel/TestMenu/TestMenu.jsx
@@ -4,13 +4,9 @@ import {
   addAction,
   addAssertion,
   addRender,
-  addAsync,
-  addReducer,
-  addActionCreator,
-  addMiddleware,
   addContexts,
   addHookUpdates,
-  addHookRender
+  addHookRender,
 } from '../../../context/testCaseActions';
 import NewTestModal from '../../NavBar/Modals/NewTestModal';
 
@@ -32,21 +28,11 @@ const TestMenu = ({ dispatchToTestCase, dispatchToMockData }) => {
   const handleAddRender = e => {
     dispatchToTestCase(addRender());
   };
-  const handleAddReducer = e => {
-    dispatchToTestCase(addReducer());
-  };
-  const handleAddActionCreator = e => {
-    dispatchToTestCase(addActionCreator());
-  };
-  const handleAddMiddleware = e => {
-    dispatchToTestCase(addMiddleware());
-  };
+
   const handleAddContext = e => {
     dispatchToTestCase(addContexts());
   };
-  const handleAddAsync = e => {
-    dispatchToTestCase(addAsync());
-  };
+
   const handleAddHookRender = e => {
     dispatchToTestCase(addHookRender());
   };
@@ -78,23 +64,12 @@ const TestMenu = ({ dispatchToTestCase, dispatchToMockData }) => {
           <button data-testid='rerenderButton' onClick={handleAddRender}>
             Rerender
           </button>
-          <button data-testid='reducerButton' onClick={handleAddReducer}>
-            Reducer
-          </button>
-          <button data-testid='asyncButton' onClick={handleAddAsync}>
-            Async Action Creator
-          </button>
-          <button data-testid='middlewareButton' onClick={handleAddMiddleware}>
-            Middleware
-          </button>
           <button data-testid='contextButton' onClick={handleAddContext}>
             Context
           </button>
-          <button data-testid='actionCreatorButton' onClick={handleAddActionCreator}>
-            Action Creator
+          <button data-testid='hookRenderButton' onClick={handleAddHookRender}>
+            Hook: Rendering
           </button>
-          <button data-testid='hookRenderButton' onClick={handleAddHookRender}>Hook: Rendering</button>
-
           <button data-testid='hookUpdatesButton' onClick={handleAddHookUpdates}>
             Hook: Updates
           </button>
