+++ src/context/reduxTestCaseActions.js
@@ -3,27 +3,27 @@ export const actionTypes = {
     UPDATE_TEST_STATEMENT: 'UPDATE_REDUX_TEST_STATEMENT',
 
     TOGGLE_REDUX: 'TOGGLE_REDUX',
-
-  
-    // ADD_ACTION: 'ADD_ACTION',
-    // DELETE_ACTION: 'DELETE_ACTION',
-    // UPDATE_ACTION: 'UPDATE_ACTION',
   
     ADD_MIDDLEWARE: 'ADD_MIDDLEWARE',
     DELETE_MIDDLEWARE: 'DELETE_MIDDLEWARE',
     UPDATE_MIDDLEWARE: 'UPDATE_MIDDLEWARE',
-  
-    // ADD_ASSERTION: 'ADD_ASSERTION',
-    // DELETE_ASSERTION: 'DELETE_ASSERTION',
-    // UPDATE_ASSERTION: 'UPDATE_ASSERTION',
-  
-    // ADD_RENDER: 'ADD_RENDER',
-    // DELETE_RENDER: 'DELETE_RENDER',
-    // UPDATE_RENDER_COMPONENT: 'UPDATE_RENDER_COMPONENT',
-  
-    // ADD_RENDER_PROP: 'ADD_RENDER_PROP',
-    // DELETE_RENDER_PROP: 'DELETE_RENDER_PROP',
-    // UPDATE_RENDER_PROP: 'UPDATE_RENDER_PROPS',
+
+    ADD_ACTIONCREATOR: 'ADD_ACTIONCREATOR',
+    DELETE_ACTIONCREATOR: 'DELETE_ACTIONCREATOR',
+    UPDATE_ACTIONCREATOR: 'UPDATE_ACTIONCREATORS',
+    
+    ADD_ASYNC: 'ADD_ASYNC',
+    DELETE_ASYNC: 'DELETE_ASYNC',
+    UPDATE_ASYNC: 'UPDATE_ASYNC',
+
+    ADD_REDUCER: 'ADD_REDUCER',
+    DELETE_REDUCER: 'DELETE_REDUCER',
+    UPDATE_REDUCER: 'UPDATE_REDUCER',
+
+    UPDATE_MIDDLEWARES_FILEPATH: 'UPDATE_MIDDLEWARES_FILEPATH',
+    UPDATE_ACTIONS_FILEPATH: 'UPDATE_ACTIONS_FILEPATH',
+    UPDATE_TYPES_FILEPATH: 'UPDATE_TYPES_FILEPATH',
+    UPDATE_REDUCERS_FILEPATH: 'UPDATE_REDUCERS_FILEPATH',
   
     CREATE_NEW_REDUX_TEST: 'CREATE_NEW_REDUX_TEST',
   };
@@ -73,98 +73,131 @@ export const actionTypes = {
     queryFunction,
     suggestions,
   });
+
+  export const addActionCreator = () => ({
+    type: actionTypes.ADD_ACTIONCREATOR,
+  });
   
-//   export const addAction = () => ({
-//     type: actionTypes.ADD_ACTION,
-//   });
+  export const deleteActionCreator = id => ({
+    type: actionTypes.DELETE_ACTIONCREATOR,
+    id,
+  });
   
-//   export const deleteAction = id => ({
-//     type: actionTypes.DELETE_ACTION,
-//     id,
-//   });
+  export const updateActionCreator = ({
+    id,
+    actionsFileName,
+    filePath,
+    typesFileName,
+    typesFilePath,
+    actionCreatorFunc,
+    actionType,
+    payloadKey,
+    payloadType,
+  }) => ({
+    type: actionTypes.UPDATE_ACTIONCREATOR,
+    id,
+    actionsFileName,
+    filePath,
+    typesFileName,
+    typesFilePath,
+    actionCreatorFunc,
+    actionType,
+    payloadKey,
+    payloadType,
+  });
+
+  export const addAsync = () => ({
+    type: actionTypes.ADD_ASYNC,
+  });
   
-//   export const updateAction = ({
-//     id,
-//     eventType,
-//     eventValue,
-//     queryVariant,
-//     querySelector,
-//     queryValue,
-//     suggestions,
-//   }) => ({
-//     type: actionTypes.UPDATE_ACTION,
-//     id,
-//     eventType,
-//     eventValue,
-//     queryVariant,
-//     querySelector,
-//     queryValue,
-//     suggestions,
-//   });
+  export const deleteAsync = id => ({
+    type: actionTypes.DELETE_ASYNC,
+    id,
+  });
   
-//   export const addAssertion = () => ({
-//     type: actionTypes.ADD_ASSERTION,
-//   });
+  export const updateAsync = ({
+    id,
+    actionsFileName,
+    filePath,
+    typesFileName,
+    typesFilePath,
+    asyncFunction,
+    method,
+    route,
+    store,
+    matcher,
+    expectedResponse,
+  }) => ({
+    type: actionTypes.UPDATE_ASYNC,
+    id,
+    actionsFileName,
+    filePath,
+    typesFileName,
+    typesFilePath,
+    asyncFunction,
+    method,
+    route,
+    store,
+    matcher,
+    expectedResponse,
+  });
   
-//   export const deleteAssertion = id => ({
-//     type: actionTypes.DELETE_ASSERTION,
-//     id,
-//   });
+  export const addReducer = () => ({
+    type: actionTypes.ADD_REDUCER,
+  });
   
-//   export const updateAssertion = ({
-//     id,
-//     queryVariant,
-//     querySelector,
-//     queryValue,
-//     isNot,
-//     matcherType,
-//     matcherValue,
-//     suggestions,
-//   }) => ({
-//     type: actionTypes.UPDATE_ASSERTION,
-//     id,
-//     queryVariant,
-//     querySelector,
-//     queryValue,
-//     isNot,
-//     matcherType,
-//     matcherValue,
-//     suggestions,
-//   });
+  export const deleteReducer = id => ({
+    type: actionTypes.DELETE_REDUCER,
+    id,
+  });
   
-//   export const addRender = () => ({
-//     type: actionTypes.ADD_RENDER,
-//   });
+  export const updateReducer = ({
+    id,
+    queryVariant,
+    querySelector,
+    queryValue,
+    typesFileName,
+    typesFilePath,
+    reducersFileName,
+    reducersFilePath,
+    matcherValue,
+  }) => ({
+    type: actionTypes.UPDATE_REDUCER,
+    id,
+    queryVariant,
+    querySelector,
+    queryValue,
+    typesFileName,
+    typesFilePath,
+    reducersFileName,
+    reducersFilePath,
+    matcherValue,
+  });
+
+  export const updateActionsFilePath = (actionsFileName, filePath) => ({
+    type: actionTypes.UPDATE_ACTIONS_FILEPATH,
+    actionsFileName,
+    filePath,
+  });
   
-//   export const deleteRender = id => ({
-//     type: actionTypes.DELETE_RENDER,
-//     id,
-//   });
+  export const updateTypesFilePath = (typesFileName, typesFilePath) => ({
+    type: actionTypes.UPDATE_TYPES_FILEPATH,
+    typesFileName,
+    typesFilePath,
+  });
   
-//   export const updateRenderComponent = (componentName, filePath) => ({
-//     type: actionTypes.UPDATE_RENDER_COMPONENT,
-//     componentName,
-//     filePath,
-//   });
+  export const updateReducersFilePath = (reducersFileName, reducersFilePath) => ({
+    type: actionTypes.UPDATE_REDUCERS_FILEPATH,
+    reducersFileName,
+    reducersFilePath,
+  });
   
-//   export const addRenderProp = renderId => ({
-//     type: actionTypes.ADD_RENDER_PROP,
-//     renderId,
-//   });
-  
-//   export const deleteRenderProp = (renderId, propId) => ({
-//     type: actionTypes.DELETE_RENDER_PROP,
-//     renderId,
-//     propId,
-//   });
-  
-//   export const updateRenderProp = (renderId, propId, propKey, propValue) => ({
-//     type: actionTypes.UPDATE_RENDER_PROP,
-//     renderId,
-//     propId,
-//     propKey,
-//     propValue,
-//   });
+  export const updateMiddlewaresFilePath = (middlewaresFileName, middlewaresFilePath) => ({
+    type: actionTypes.UPDATE_MIDDLEWARES_FILEPATH,
+    middlewaresFileName,
+    middlewaresFilePath,
+  });
+
   export const createNewReduxTest = () => ({
     type: actionTypes.CREATE_NEW_REDUX_TEST,
   });
